mobilebackend/requirements.txt
APScheduler==3.10.1
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
flake8==3.8.4
fpdf==1.7.2
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.6.1
pdfkit==1.0.0
Pillow==9.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.5.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
sqlparse==0.3.1
twilio==7.17.0
tzdata==2022.6
tzlocal==4.3
urllib3==1.26.15
whitenoise==6.2.0
et-xmlfile==1.1.0
openpyxl==3.1.2
